[wasm64] Use `int` over `long` for time_t, etc. NFC
This doesn't change the size of the type for wasm32 but helps with
wasm64 since it makes these types consistent between the two.

Split out from #16922

+++ system/lib/libc/musl/src/time/clock.c
@@ -8,8 +8,8 @@ clock_t clock()
 	if (__clock_gettime(CLOCK_PROCESS_CPUTIME_ID, &ts))
 		return -1;
 
-	if (ts.tv_sec > LONG_MAX/1000000
-	 || ts.tv_nsec/1000 > LONG_MAX-1000000*ts.tv_sec)
+	if (ts.tv_sec > INT_MAX/1000000
+	 || ts.tv_nsec/1000 > INT_MAX-1000000*ts.tv_sec)
 		return -1;
 
 	return ts.tv_sec*1000000 + ts.tv_nsec/1000;
